requirement-google.txt
langgraph>=0.0.28
python-dotenv>=1.0.1
google-cloud-speech>=2.27.0
